+++ src/components/pages/01-GestioneCommesse.js
@@ -1,8 +1,9 @@
 import React, { useState, useEffect } from "react";
 import "../style.css";
 import CommessaCrea from "../CommessaCrea"; 
-import logo from"../assets/unitech-packaging.png";
-
+import logo from "../assets/Animation - 1738249246846.gif";
+import { ToastContainer, toast } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
 
 import {
   fetchCommesse,
@@ -16,28 +17,26 @@ function GestioneCommesse() {
   const [commesse, setCommesse] = useState([]);
   const [reparti, setReparti] = useState([]);
   const [attivita, setAttivita] = useState([]);
-  const [selectedCommessa, setSelectedCommessa] = useState(null); 
-  const [isEditing, setIsEditing] = useState(false); 
-  const [showPopup, setShowPopup] = useState(false); 
+  const [selectedCommessa, setSelectedCommessa] = useState(null);
+  const [isEditing, setIsEditing] = useState(false);
+  const [showPopup, setShowPopup] = useState(false);
   const [selezioniAttivita, setSelezioniAttivita] = useState({});
   const [editId, setEditId] = useState(null);
-   const [loading, setLoading] = useState(false);
+  const [loading, setLoading] = useState(false);
+  const [deleteLoadingId, setDeleteLoadingId] = useState(null);
   const [showFilters, setShowFilters] = useState(false);
   const [clienteFilter, setClienteFilter] = useState("");
   const [tipoMacchinaFilter, setTipoMacchinaFilter] = useState("");
   const [commessaFilter, setCommessaFilter] = useState("");
-  const [showClienteSuggestions, setShowClienteSuggestions] = useState(false);
-  const [showTipoMacchinaSuggestions, setShowTipoMacchinaSuggestions] = useState(false);
-  const [showCommessaSuggestions, setShowCommessaSuggestions] = useState(false);
-  const [statiCommessa, setStatiCommessa] = useState([]); 
+  const [statiCommessa, setStatiCommessa] = useState([]);
 
   useEffect(() => {
     loadData();
   }, []);
 
   const loadData = async () => {
+    setLoading(true);
     try {
-      setLoading(true);
       const [commesseData, repartiData, attivitaData, statiData] = await Promise.all([
         fetchCommesse(),
         fetchReparti(),
@@ -50,44 +49,47 @@ function GestioneCommesse() {
       setStatiCommessa(statiData);
     } catch (error) {
       console.error("Errore durante il caricamento dei dati:", error);
+      toast.error("Errore durante il caricamento dei dati.");
     } finally {
       setLoading(false);
     }
   };
 
-  // Funzione per aprire il pop-up in modalità creazione
   const handleCreateNewCommessa = () => {
-    setIsEditing(false); 
-    setSelectedCommessa(null); 
-    setShowPopup(true); 
+    setIsEditing(false);
+    setSelectedCommessa(null);
+    setShowPopup(true);
   };
 
-  // Funzione per aprire il pop-up in modalità modifica
   const handleEditCommessa = (commessa) => {
-    setIsEditing(true); 
-    setSelectedCommessa(commessa); 
-    setEditId(commessa.commessa_id); 
-    setShowPopup(true); 
+    setIsEditing(true);
+    setSelectedCommessa(commessa);
+    setEditId(commessa.commessa_id);
+    setShowPopup(true);
   };
 
-  // Funzione per chiudere il pop-up
   const handleClosePopup = () => {
-   setShowPopup(false); 
-    setSelectedCommessa(null); 
+    setShowPopup(false);
+    setSelectedCommessa(null);
+    setEditId(null);
   };
 
-  // Funzione per eliminare una commessa
   const handleDelete = async (commessaId) => {
-    try {
-      await deleteCommessa(commessaId);
-      await loadData(); // Ricarica i dati dopo l'eliminazione
-    } catch (error) {
-      console.error("Errore durante l'eliminazione della commessa:", error);
+    if (window.confirm("Sei sicuro di voler eliminare questa commessa?")) {
+      setDeleteLoadingId(commessaId);
+      try {
+        await deleteCommessa(commessaId);
+        toast.success("Commessa eliminata con successo!");
+        await loadData();
+      } catch (error) {
+        console.error("Errore durante l'eliminazione della commessa:", error);
+        toast.error("Errore durante l'eliminazione della commessa.");
+      } finally {
+        setDeleteLoadingId(null);
+      }
     }
   };
 
-
-  // Funzione per applicare i filtri
   const applyFilters = () => {
     return commesse.filter((commessa) => {
       return (
@@ -98,123 +100,56 @@ function GestioneCommesse() {
     });
   };
 
-  // Funzioni per gestire i cambiamenti dei filtri
-  const handleCommessaChange = (e) => {
-    setCommessaFilter(e.target.value);
-    setShowCommessaSuggestions(true);
-  };
-
-  const handleClienteChange = (e) => {
-    setClienteFilter(e.target.value);
-    setShowClienteSuggestions(true);
-  };
-
-  const handleTipoMacchinaChange = (e) => {
-    setTipoMacchinaFilter(e.target.value);
-    setShowTipoMacchinaSuggestions(true);
-  };
-
-  const closeSuggestions = () => {
-    setShowClienteSuggestions(false);
-    setShowTipoMacchinaSuggestions(false);
-    setShowCommessaSuggestions(false);
-  };
-
-  const toggleFilters = () => {
-    setShowFilters((prev) => !prev);
-  };
-
   const getStatoNome = (id) => {
-    const stato = statiCommessa.find(stato => stato.id === id);
-    return stato ? stato.nome_stato : "Non assegnato"; 
+    const stato = statiCommessa.find((stato) => stato.id === id);
+    return stato ? stato.nome_stato : "Non assegnato";
   };
 
-  
   return (
-    <div className="container" onClick={closeSuggestions}>
+    <div className="container">
+      <ToastContainer position="top-right" autoClose={3000} hideProgressBar />
       {loading && (
         <div className="loading-overlay">
-            <img src={logo} alt="Logo"  className="logo-spinner"/>
+          <img src={logo} alt="Logo" className="logo-spinner" />
         </div>
       )}
 
       <div className="header">
-      <h1>Commesse</h1>
-      </div>
-      <button onClick={handleCreateNewCommessa} className="btn btn-primary create-activity-btn">
-        Crea Nuova Commessa
-      </button>
-        <button onClick={toggleFilters} className="btn btn-filter">
+        <h1>Gestione Commesse</h1>
+        <button onClick={handleCreateNewCommessa} className="btn btn-primary">
+          Crea Nuova Commessa
+        </button>
+        <button onClick={() => setShowFilters(!showFilters)} className="btn btn-filter">
           {showFilters ? "Nascondi Filtri" : "Mostra Filtri"}
-        </button>     
-        {showFilters && (
-      <div className="filters">
-        <div className="filter-group">
-        <input
-          type="text"
-          placeholder="Filtra per Numero Commessa"
-          value={commessaFilter}
-          onChange={handleCommessaChange}
-           className="input-field"
-        />
-        {showCommessaSuggestions && (
-          <ul className="suggestions-list">
-            {commesse
-              .filter((commessa) => commessa.numero_commessa.includes(commessaFilter))
-              .map((commessa, index) => (
-                <li key={index} onClick={() => setCommessaFilter(commessa.numero_commessa)}>
-                  {commessa.numero_commessa}
-                </li>
-              ))}
-          </ul>
-        )}
+        </button>
       </div>
 
-      <div className="filter-group">
-        <input
-          type="text"
-          placeholder="Filtra per Cliente"
-          value={clienteFilter}
-          onChange={handleClienteChange}
-          className="input-field"
-        />
-        {showClienteSuggestions && (
-          <ul className="suggestions-list">
-            {commesse
-              .map((commessa) => commessa.cliente)
-              .filter((cliente, index, self) => self.indexOf(cliente) === index)
-              .map((cliente, index) => (
-                <li key={index} onClick={() => setClienteFilter(cliente)}>
-                  {cliente}
-                </li>
-              ))}
-          </ul>
-        )}
-      </div>
+      {showFilters && (
+        <div className="filters">
+          <input
+            type="text"
+            placeholder="Filtra per Numero Commessa"
+            value={commessaFilter}
+            onChange={(e) => setCommessaFilter(e.target.value)}
+            className="input-field"
+          />
+          <input
+            type="text"
+            placeholder="Filtra per Cliente"
+            value={clienteFilter}
+            onChange={(e) => setClienteFilter(e.target.value)}
+            className="input-field"
+          />
+          <input
+            type="text"
+            placeholder="Filtra per Tipo Macchina"
+            value={tipoMacchinaFilter}
+            onChange={(e) => setTipoMacchinaFilter(e.target.value)}
+            className="input-field"
+          />
+        </div>
+      )}
 
-      <div className="filter-group">
-        <input
-          type="text"
-          placeholder="Filtra per Tipo Macchina"
-          value={tipoMacchinaFilter}
-          onChange={handleTipoMacchinaChange}
-          className="input-field"
-        />
-        {showTipoMacchinaSuggestions && (
-          <ul className="suggestions-list">
-            {commesse
-              .map((commessa) => commessa.tipo_macchina)
-              .filter((tipo, index, self) => self.indexOf(tipo) === index)
-              .map((tipo, index) => (
-                <li key={index} onClick={() => setTipoMacchinaFilter(tipo)}>
-                  {tipo}
-                </li>
-              ))}
-          </ul>
-        )}
-      </div>
-      </div>
-          )}
       <table>
         <thead>
           <tr>
@@ -228,26 +163,24 @@ function GestioneCommesse() {
           </tr>
         </thead>
         <tbody>
-        {applyFilters().map((commessa) => (
+          {applyFilters().map((commessa) => (
             <tr key={commessa.id}>
               <td>{commessa.numero_commessa}</td>
               <td>{commessa.tipo_macchina}</td>
               <td>{commessa.cliente}</td>
               <td>{new Date(commessa.data_consegna).toLocaleDateString()}</td>
-              <td> {commessa.data_FAT? new Date(commessa.data_FAT).toLocaleDateString(): "Non specificata"}</td>
-              <td>{getStatoNome(commessa.stato)}</td> 
+              <td>{commessa.data_FAT ? new Date(commessa.data_FAT).toLocaleDateString() : "Non specificata"}</td>
+              <td>{getStatoNome(commessa.stato)}</td>
               <td>
-                <button
-                  className="btn btn-warning"
-                  onClick={() => handleEditCommessa(commessa)}
-                >
+                <button className="btn btn-warning" onClick={() => handleEditCommessa(commessa)}>
                   Modifica
                 </button>
                 <button
                   className="btn btn-danger"
                   onClick={() => handleDelete(commessa.commessa_id)}
+                  disabled={deleteLoadingId === commessa.commessa_id}
                 >
-                  Elimina
+                  {deleteLoadingId === commessa.commessa_id ? "Eliminazione..." : "Elimina"}
                 </button>
               </td>
             </tr>
@@ -255,19 +188,18 @@ function GestioneCommesse() {
         </tbody>
       </table>
 
-      {/* Mostra il pop-up se c'è una commessa selezionata o se non siamo in modalità di creazione */}
       {showPopup && (
         <CommessaCrea
-        commessa={selectedCommessa}
-        onClose={handleClosePopup}
-        isEditing={isEditing}
-        reparti={reparti}
-        attivita={attivita}
-        selezioniAttivita={selezioniAttivita}
-        setSelezioniAttivita={setSelezioniAttivita}
-        fetchCommesse={fetchCommesse}
-        editId={editId}
-        stato={statiCommessa}
+          commessa={selectedCommessa}
+          onClose={handleClosePopup}
+          isEditing={isEditing}
+          reparti={reparti}
+          attivita={attivita}
+          selezioniAttivita={selezioniAttivita}
+          setSelezioniAttivita={setSelezioniAttivita}
+          fetchCommesse={loadData}
+          editId={editId}
+          stato={statiCommessa}
         />
       )}
     </div>
